[A] ADDED EventsMixin Cog [F] naming fix for channels enum
